% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MclustBounded.R
\name{mclustAddons-internal}
\alias{mclustAddons-internal}
\alias{as.MclustBounded}
\alias{as.MclustBounded.default}
\alias{as.MclustBounded.densityMclustBounded}
\alias{as.densityMclustBounded}
\alias{as.densityMclustBounded.default}
\alias{as.densityMclustBounded.MclustBounded}
\title{Internal \pkg{mclustAddons} functions}
\usage{
as.MclustBounded(x, ...)

\method{as.MclustBounded}{default}(x, ...)

\method{as.MclustBounded}{densityMclustBounded}(x, ...)

as.densityMclustBounded(x, ...)

\method{as.densityMclustBounded}{default}(x, ...)

\method{as.densityMclustBounded}{MclustBounded}(x, ...)
}
\arguments{
\item{x}{An object of class specific for the method.}

\item{\dots}{Further arguments passed to or from other methods.}
}
\description{
Internal functions not intended to be called directly by users.
}
