% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MclustBounded.R
\name{plot.MclustBounded}
\alias{plot.MclustBounded}
\title{Plotting method for model-based clustering of bounded data}
\usage{
\method{plot}{MclustBounded}(x, what = c("BIC", "classification", "uncertainty"), dimens = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{'MclustBounded'}.}

\item{what}{A string specifying the type of graph requested.
Available choices are:
\itemize{
\item \code{"BIC"} Plot of BIC values used for choosing the number of clusters.
\item \code{"classification"} Plot showing the clustering. For data in more than two
dimensions, a scatterplot of pairwise coordinate projections using the
specified \code{dimens} is produced.
\item \code{"uncertainty"} Plot of classification uncertainty. For data in more than
two dimensions, a scatterplot of pairwise coordinate projections using the
specified \code{dimens} is produced.
}}

\item{dimens}{A vector of integers specifying the dimensions of the coordinate
projections.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
No return value, called for side effects.
}
\description{
Plotting method for model-based clustering of bounded data
}
