% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mclustAddons-package.R
\docType{data}
\name{racial}
\alias{racial}
\title{Racial data}
\format{
A data frame with the following variables:
\describe{
\item{District}{School district.}
\item{PropWhite}{Proportion of white student enrolled.}
}
}
\source{
Simonoff, S.J. (1996) Smoothing Methods in Statistics, Springer-Verlag,
New York, p. 52
}
\description{
Proportion of white student enrollment in 56 school districts in Nassau
County (Long Island, New York), for the 1992-1993 school year.
}
\keyword{datasets}
