% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mclustAddons-package.R
\docType{data}
\name{suicide}
\alias{suicide}
\title{Suicide data}
\format{
A vector of containing the lengths (days) of 86 spells of
psychiatric treatment undergone by patients used as controls in a study of
suicide risks.
}
\source{
Silverman, B. W. (1986) Density Estimation, Chapman & Hall, Tab 2.1.
}
\description{
Lengths of treatment spells (in days) of control patients in suicide study.
}
\keyword{datasets}
