% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMM_finance.R
\name{VaR}
\alias{VaR}
\alias{ES}
\title{Financial risk measures}
\usage{
VaR(object, ...)

ES(object, ...)
}
\arguments{
\item{object}{An object of class specific for the method.}

\item{\dots}{Further arguments passed to or from other methods.}
}
\description{
Generic functions for computing Value-at-Risk (VaR) and Expected
Shortfall (ES).
}
